+++ apps/client/src/app/pages/markets/markets-page.module.ts
@@ -1,4 +1,4 @@
-import { GfHomeMarketModule } from '@ghostfolio/client/components/home-market/home-market.module';
+import { HomeMarketComponent } from '@ghostfolio/client/components/home-market/home-market.component';
 
 import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
@@ -8,7 +8,7 @@ import { MarketsPageComponent } from './markets-page.component';
 
 @NgModule({
   declarations: [MarketsPageComponent],
-  imports: [CommonModule, GfHomeMarketModule, MarketsPageRoutingModule],
+  imports: [CommonModule, HomeMarketComponent, MarketsPageRoutingModule],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class MarketsPageModule {}